+++ pyobjc-core/PyObjCTest/test_python_method.py
@@ -3,6 +3,7 @@ import gc
 from PyObjCTools.TestSupport import TestCase
 
 NSObject = objc.lookUpClass("NSObject")
+NSMutableArray = objc.lookUpClass("NSMutableArray")
 
 
 class TestPythonMethod(TestCase):
@@ -74,6 +75,15 @@ class TestPythonMethod(TestCase):
 
         o = Foo()
         self.assertEqual(o.args(b=1, a=2), (2, 1))
+
+    def test_create_alias(self):
+        NSMutableArray.ocInitWithArray = objc.python_method(
+            NSMutableArray.initWithArray_
+        )
+
+        o = NSMutableArray.alloc().ocInitWithArray([1, 2])
+        self.assertIsInstance(o, NSMutableArray)
+        self.assertEqual(o, [1, 2])
 
     def test_gc(self):
         deallocated = False